backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
cryptography==42.0.8
mysql-connector-python==8.2.0
alembic==1.12.1 
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib 